requirements.txt
pandas
pytz
skyfield
swisseph
Flask
requests
openpyxl
python-dotenv
scikit-learn
